Enable "Respond via SMS" option.
In current implementation, when "Ask First" is enabled for SMS
preferred setting, then "Respond via SMS" option is shown for
MT call on one sub and not shown for MT call on other sub leading
to inconsistency.
As part of this fix, enable "Respond via SMS" option for both subs
regardless of what is SMS preferred setting.

Change-Id: Idc666c151b8519a350ec56bf2ad756c21d43d223
CRs-Fixed: 972040

 Conflicts:
	src/com/android/server/telecom/ParcelableCallUtils.java

Signed-off-by: NoelMacwan <yulespencer@gmail.com>